requirements.txt
asgiref==3.8.1
Django==5.1.7
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
pymysql
